fix for delete errors in BASE-755

+++ src/main/kotlin/com/bitclave/node/repository/search/interaction/PostgresOfferInteractionRepositoryImpl.kt
@@ -56,6 +56,8 @@ class PostgresOfferInteractionRepositoryImpl(
     }
 
     override fun delete(ids: List<Long>): Int {
+        if (ids.size == 0)    return 0
+
         repository.deleteEventsByIdIn(ids)
         return repository.deleteByIdIn(ids)
     }
